refactor: Move existing implementation to /legacy dir
