+++ packages/json-rpc-engine/src/legacy/createAsyncMiddleware.test.ts
@@ -1,6 +1,6 @@
 import { assertIsJsonRpcSuccess } from '@metamask/utils';
 
-import { JsonRpcEngine, createAsyncMiddleware } from '.';
+import { JsonRpcEngine, createAsyncMiddleware } from '..';
 
 const jsonrpc = '2.0' as const;
 
@@ -37,7 +37,7 @@ describe('createAsyncMiddleware', () => {
         await next();
         expect(response.result).toBe(1234);
         // override value
-        response.result = 42; // eslint-disable-line require-atomic-updates
+        response.result = 42;
       }),
     );
