+++ packages/json-rpc-engine/src/legacy/createScaffoldMiddleware.ts
